refactor: fetch benchmark scores concurrently via async sources
* Convert benchmark fetchers to async
* Run benchmark fetchers in parallel
* Move all fetchers to their respective modules, move reused utilities/types
* Move _current_version out of cli.py into whichllm/utils.py
  so it can be used for user-agent formation
* Don't quietly fall back to stale data in aa_index and livebench
