+++ src/whichllm/models/benchmark_sources/__init__.py
@@ -9,9 +9,19 @@ returns malformed data, they log a warning and return an empty dict so the
 main benchmark merge pipeline does not abort.
 """
 
-from whichllm.models.benchmark_sources.aa_index import fetch_aa_index_scores
+from whichllm.models.benchmark_sources.aa_index import (
+    fetch_aa_index_scores,
+    get_aa_curated_fallback,
+)
 from whichllm.models.benchmark_sources.aider import fetch_aider_polyglot_scores
-from whichllm.models.benchmark_sources.livebench import fetch_livebench_scores
+from whichllm.models.benchmark_sources.chatbot_arena import fetch_arena_scores
+from whichllm.models.benchmark_sources.livebench import (
+    fetch_livebench_scores,
+    get_livebench_curated_fallback,
+)
+from whichllm.models.benchmark_sources.open_llm_leaderboard import (
+    fetch_leaderboard_with_fallback,
+)
 from whichllm.models.benchmark_sources.vision import fetch_vision_scores
 
 # Newest curated-fallback date across all sources. Live scrapes are merged
@@ -26,6 +36,10 @@ __all__ = [
     "BENCHMARK_SNAPSHOT",
     "fetch_aa_index_scores",
     "fetch_aider_polyglot_scores",
+    "fetch_arena_scores",
+    "fetch_leaderboard_with_fallback",
     "fetch_livebench_scores",
     "fetch_vision_scores",
+    "get_aa_curated_fallback",
+    "get_livebench_curated_fallback",
 ]
