+++ src/whichllm/models/benchmark_sources/constants.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+import re
+
+_NEXT_DATA_RE = re.compile(
+    r'<script id="__NEXT_DATA__"[^>]*>(?P<json>.*?)</script>', re.DOTALL
+)
