+++ src/whichllm/models/benchmark_sources/types.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+
+class ExtractionFailed(Exception):
+    pass
